requirements.txt
nbformat>=5.0.0
nbconvert>=6.0.0
jupyter>=1.0.0
ipykernel>=6.0.0
python-dotenv

yfinance
scipy
numba
backtesting
seaborn
polars
plotly
hftbacktest

#hft_strat ##not public
hmmlearn
yfinance
statsmodels
#algo_deploy ##not in PyPi
pandas_ta
hurst
hmmlearn
binance_historical_data
binance_bulk_downloader
pandas_ta
#talib
arch
#trades_from_signal
indicator
vnstock
#pypfopt
PyPortfolioOpt
filterpy
#google-colab
mplfinance
quantstats
#fetch_data_binance